requirements.txt
# Core SDK
google-adk>=0.6.0
# Google Cloud client & Vertex AI
google-cloud-aiplatform>=1.50.0
# Research tooling
arxiv==2.1.3
semanticscholar==0.6.0
feedparser==6.0.11   # (secondary RSS ingest)
# PDF/LATEX
jinja2>=3.1.3
reportlab>=4.1.0
# Web UI
gradio>=4.0.0
# Testing
pytest>=7.0.0
pytest-mock>=3.10.0
# Misc
python-dotenv>=1.0.1
tqdm>=4.65.0      # Progress bars
colorama>=0.4.6   # Colored terminal output 